Link between elements + upgrade script for models
Elements can be linked between them using a modal. Elements that are
sent to the trash keep the information of the elements linked to them,
but the elements they are linked to lose the link (can't link an element
that has been trashed)
A basic script to update elements whenever models are changed has been
implemented. Of course the actions of the script need to be updated
according to what has changed in the new version.